STN-botv3/requirements.txt
# STN-bot v3 Dependencies
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.17.0
